Split and link `textlib`
