+++ config/GALE01/splits.txt
@@ -4533,16 +4533,23 @@ melee/if/textdraw.c:
 	.rodata     start:0x803B8790 end:0x803B87A8
 	.data       start:0x803FDC20 end:0x803FDC80
 	.bss        start:0x804A1FD8 end:0x804A2688
-	.sbss       start:0x804D6E18 end:0x804D6E50
+	.sbss       start:0x804D6E18 end:0x804D6E40
 	.sdata2     start:0x804DDC58 end:0x804DDC88
 
 melee/if/textlib.c:
-	.text       start:0x80302834 end:0x80304870
+	.text       start:0x80302834 end:0x80302E00
+	.data       start:0x803FDC80 end:0x803FDCA8
+	.sdata      start:0x804D5A00 end:0x804D5A08
+	.sdata2     start:0x804DDC88 end:0x804DDCA8
+
+melee/if/textlib_1.c:
+	.text       start:0x80302E00 end:0x80304870
 	.rodata     start:0x803B87A8 end:0x803B8844
-	.data       start:0x803FDC80 end:0x803FDD18
+	.data       start:0x803FDCA8 end:0x803FDD18
 	.bss        start:0x804A2688 end:0x804A26B8
-	.sdata      start:0x804D5A00 end:0x804D5A40
-	.sdata2     start:0x804DDC88 end:0x804DDCC8
+	.sdata      start:0x804D5A08 end:0x804D5A40
+	.sbss       start:0x804D6E40 end:0x804D6E50
+	.sdata2     start:0x804DDCA8 end:0x804DDCC8
 
 melee/ty/toy.c:
 	.text       start:0x80304870 end:0x80312834
