+++ config/GALE01/symbols.txt
@@ -20147,9 +20147,9 @@ it_803B8730 = .rodata:0x803B8730; // type:object size:0xC scope:global data:4byt
 @208 = .rodata:0x803B8780; // type:object size:0x10 scope:local data:4byte
 @229 = .rodata:0x803B8790; // type:object size:0xC scope:local data:4byte
 @230 = .rodata:0x803B879C; // type:object size:0xC scope:local data:4byte
-un_803B87A8 = .rodata:0x803B87A8; // type:object size:0x34 scope:global data:short
-un_803B87DC = .rodata:0x803B87DC; // type:object size:0x34 scope:global data:short
-un_803B8810 = .rodata:0x803B8810; // type:object size:0x34 scope:global data:short
+@741 = .rodata:0x803B87A8; // type:object size:0x34 scope:local data:short
+@754 = .rodata:0x803B87DC; // type:object size:0x34 scope:local data:short
+@767 = .rodata:0x803B8810; // type:object size:0x34 scope:local data:short
 _Toy_803B8844 = .rodata:0x803B8844; // type:object size:0x14 scope:local data:4byte
 _Toy_803B8858 = .rodata:0x803B8858; // type:object size:0xC scope:local data:4byte
 _Toy_803B8864 = .rodata:0x803B8864; // type:object size:0x38 scope:local data:4byte
@@ -24769,11 +24769,11 @@ devtext_eyepos = .data:0x803FDC20; // type:object size:0x14 scope:local
 ...data.0 = .data:0x803FDC20; // type:label scope:local
 devtext_interest = .data:0x803FDC34; // type:object size:0x14 scope:local
 devtext_CObjDesc = .data:0x803FDC48; // type:object size:0x38 scope:local
-un_803FDC80 = .data:0x803FDC80; // type:object size:0x18 scope:global data:string
-un_803FDC98 = .data:0x803FDC98; // type:object size:0xA scope:global data:string
-jtbl_803FDCA8 = .data:0x803FDCA8; // type:object size:0x24 scope:local
-jtbl_803FDCCC = .data:0x803FDCCC; // type:object size:0x24 scope:local
-jtbl_803FDCF0 = .data:0x803FDCF0; // type:object size:0x24 scope:local
+@205 = .data:0x803FDC80; // type:object size:0x18 scope:local data:string
+@206 = .data:0x803FDC98; // type:object size:0xA scope:local data:string
+@332 = .data:0x803FDCA8; // type:object size:0x24 scope:local
+@377 = .data:0x803FDCCC; // type:object size:0x24 scope:local
+@422 = .data:0x803FDCF0; // type:object size:0x24 scope:local
 _Toy_str_TyLight_dat = .data:0x803FDD18; // type:object size:0xC scope:local data:string
 _Toy_803FDD24 = .data:0x803FDD24; // type:object size:0x1F scope:local data:string
 _Toy_803FDD44 = .data:0x803FDD44; // type:object size:0x1F scope:local data:string
@@ -27953,16 +27953,16 @@ un_804D59E0 = .sdata:0x804D59E0; // type:object size:0x6 scope:local data:string
 un_804D59E8 = .sdata:0x804D59E8; // type:object size:0x6 scope:local data:string
 un_804D59F0 = .sdata:0x804D59F0; // type:object size:0x7 scope:local data:string
 un_804D59F8 = .sdata:0x804D59F8; // type:object size:0x7 scope:local data:string
-un_804D5A00 = .sdata:0x804D5A00; // type:object size:0x2 scope:local data:string
+@207 = .sdata:0x804D5A00; // type:object size:0x2 scope:local data:string
 un_804D5A08 = .sdata:0x804D5A08; // type:object size:0x4 scope:local data:4byte
 un_804D5A0C = .sdata:0x804D5A0C; // type:object size:0x4 scope:local data:4byte
 un_804D5A10 = .sdata:0x804D5A10; // type:object size:0x4 scope:local data:4byte
 un_804D5A14 = .sdata:0x804D5A14; // type:object size:0x4 scope:local data:4byte
-un_804D5A18 = .sdata:0x804D5A18; // type:object size:0x3 scope:local data:string
-un_804D5A1C = .sdata:0x804D5A1C; // type:object size:0x5 scope:local data:string
-un_804D5A24 = .sdata:0x804D5A24; // type:object size:0x5 scope:local data:string
-un_804D5A2C = .sdata:0x804D5A2C; // type:object size:0x5 scope:local data:string
-un_804D5A34 = .sdata:0x804D5A34; // type:object size:0x6 scope:local data:string
+@327 = .sdata:0x804D5A18; // type:object size:0x3 scope:local data:string
+@328 = .sdata:0x804D5A1C; // type:object size:0x5 scope:local data:string
+@329 = .sdata:0x804D5A24; // type:object size:0x5 scope:local data:string
+@330 = .sdata:0x804D5A2C; // type:object size:0x5 scope:local data:string
+@331 = .sdata:0x804D5A34; // type:object size:0x6 scope:local data:string
 _Toy_color_E2E2E2FF = .sdata:0x804D5A40; // type:object size:0x4 scope:local data:4byte
 _Toy_color_FF8020FF = .sdata:0x804D5A44; // type:object size:0x4 scope:local data:4byte
 _Toy_str___FILE__ = .sdata:0x804D5A48; // type:object size:0x6 scope:local data:string
@@ -34404,18 +34404,18 @@ un_804DDC50 = .sdata2:0x804DDC50; // type:object size:0x4 scope:global align:4 d
 @273 = .sdata2:0x804DDC74; // type:object size:0x4 scope:local align:4 data:float
 @275 = .sdata2:0x804DDC78; // type:object size:0x8 scope:local align:8 data:double
 @278 = .sdata2:0x804DDC80; // type:object size:0x8 scope:local align:8 data:double
-un_804DDC88 = .sdata2:0x804DDC88; // type:object size:0x4 scope:global data:4byte
-un_804DDC8C = .sdata2:0x804DDC8C; // type:object size:0x4 scope:global data:4byte
-un_804DDC90 = .sdata2:0x804DDC90; // type:object size:0x4 scope:global data:4byte
-un_804DDC94 = .sdata2:0x804DDC94; // type:object size:0x4 scope:global data:4byte
-un_804DDC98 = .sdata2:0x804DDC98; // type:object size:0x4 scope:global data:4byte
-un_804DDC9C = .sdata2:0x804DDC9C; // type:object size:0x4 scope:global align:4 data:float
-un_804DDCA0 = .sdata2:0x804DDCA0; // type:object size:0x4 scope:global align:4 data:float
-un_804DDCA8 = .sdata2:0x804DDCA8; // type:object size:0x4 scope:global align:4 data:float
-un_804DDCB0 = .sdata2:0x804DDCB0; // type:object size:0x8 scope:global align:8 data:double
-un_804DDCB8 = .sdata2:0x804DDCB8; // type:object size:0x4 scope:global align:4 data:float
-un_804DDCBC = .sdata2:0x804DDCBC; // type:object size:0x4 scope:global align:4 data:float
-un_804DDCC0 = .sdata2:0x804DDCC0; // type:object size:0x8 scope:global align:8 data:double
+cyan$178 = .sdata2:0x804DDC88; // type:object size:0x4 scope:local data:4byte
+white$192 = .sdata2:0x804DDC8C; // type:object size:0x4 scope:local data:4byte
+red$193 = .sdata2:0x804DDC90; // type:object size:0x4 scope:local data:4byte
+green$194 = .sdata2:0x804DDC94; // type:object size:0x4 scope:local data:4byte
+blue$195 = .sdata2:0x804DDC98; // type:object size:0x4 scope:local data:4byte
+@208 = .sdata2:0x804DDC9C; // type:object size:0x4 scope:local align:4 data:float
+@209 = .sdata2:0x804DDCA0; // type:object size:0x4 scope:local align:4 data:float
+@376 = .sdata2:0x804DDCA8; // type:object size:0x4 scope:local align:4 data:float
+@379 = .sdata2:0x804DDCB0; // type:object size:0x8 scope:local align:8 data:double
+@656 = .sdata2:0x804DDCB8; // type:object size:0x4 scope:local align:4 data:float
+@657 = .sdata2:0x804DDCBC; // type:object size:0x4 scope:local align:4 data:float
+@673 = .sdata2:0x804DDCC0; // type:object size:0x8 scope:local align:8 data:double
 _Toy_f32_100 = .sdata2:0x804DDCC8; // type:object size:0x4 scope:local align:4 data:float
 _Toy_f64_S32_TO_F32 = .sdata2:0x804DDCD0; // type:object size:0x8 scope:local align:8 data:double
 _Toy_f32_0 = .sdata2:0x804DDCD8; // type:object size:0x4 scope:local align:4 data:float
