+++ configure.py
@@ -1530,7 +1530,8 @@ config.libs = [
             Object(Matching, "melee/if/if_2FF2.c"),
             Object(NonMatching, "melee/if/soundtest.c"),
             Object(Matching, "melee/if/textdraw.c"),
-            Object(NonMatching, "melee/if/textlib.c"),
+            Object(Matching, "melee/if/textlib.c"),
+            Object(Matching, "melee/if/textlib_1.c"),
         ],
     ),
     MeleeLib(
